fix: refactored IntegrationForm, made header-based info if no connection to docker
