+++ src/components/IntegrationsView/IntegrationDocker/IntegrationDocker.module.css
@@ -1,0 +1,4 @@
+.DockerIcon {
+  max-width: 35px;
+  object-fit: cover;
+}
