+++ src/components/IntegrationsView/IntegrationForm/IntegrationForm.tsx
@@ -17,12 +17,8 @@ import { IntegrationDocker } from "../IntegrationDocker";
 import { SmartLink } from "../../SmartLink";
 import { renderIntegrationFormField } from "../../../features/Integrations/renderIntegrationFormField";
 import { IntegrationAvailability } from "./IntegrationAvailability";
-import { toPascalCase } from "../../../utils/toPascalCase";
-import { iconMap } from "../icons/iconMap";
 import { IntegrationDeletePopover } from "../IntegrationDeletePopover";
 import { debugIntegrations } from "../../../debugConfig";
-import { selectThemeMode } from "../../../features/Config/configSlice";
-import { useAppSelector } from "../../../hooks";
 import type { ToolParameterEntity } from "../../../services/refact";
 import {
   areAllFieldsBoolean,
@@ -76,11 +72,6 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
   const [areExtraFieldsRevealed, setAreExtraFieldsRevealed] = useState(false);
 
   const { integration } = useGetIntegrationDataByPathQuery(integrationPath);
-
-  const theme = useAppSelector(selectThemeMode);
-  const icons = iconMap(
-    theme ? (theme === "inherit" ? "light" : theme) : "light",
-  );
 
   const handleAvailabilityChange = useCallback(
     (fieldName: string, value: boolean) => {
@@ -345,16 +336,6 @@ export const IntegrationForm: FC<IntegrationFormProps> = ({
         )}
       {integration.data.integr_schema.docker && (
         <Flex mt="6" direction="column" align="start" gap="5">
-          <Flex gap="2" align="center" justify="center" width="100%">
-            <img
-              src={icons.docker}
-              className={styles.DockerIcon}
-              alt={integration.data.integr_name}
-            />
-            <Heading as="h3" align="left">
-              {toPascalCase(integration.data.integr_name)} Containers
-            </Heading>
-          </Flex>
           <IntegrationDocker
             dockerData={integration.data.integr_schema.docker}
             integrationName={integration.data.integr_name}
